wayland: Add window hint to render as shell layer (for example, desktop background)
This adds a wayland specific hint `GLFW_WAYLAND_SHELL_LAYER` to make it
possible to render to one of wayland's full screen layers using the
[wlr-layer-shell-unstable-v1](https://wayland.app/protocols/wlr-layer-shell-unstable-v1) wayland
interface. This can be used to, for example, render to the desktop
background, or as an overlay.

It can take any of the values from `zwlr_layer_shell_v1::layer`, or `-1`
(default) to create an application window.