Improve init and termination with counting
Counts additional calls to glfwInit and waits with termination until the same
amount of calls to glfwTerminate is done.
For normal users nothing should change.
For libraries/plug-ins they can now call glfwInit and glfwTerminate additional
times as long as the calls are balanced. This avoids the problem that they
would terminate GLFW completely when the main application is still using it.