Cocoa: Make sure NSWindowStyleMaskResizable is masked when entering fullscreen.
On macOS Catalina and earlier, not having the resizable bit cleared in
NSWindowStyleMask in fullscreen leads to windows minimising when clicked
anywhere in the content area.

TODO: This bug doesn't seem to occur on macOS Big Sur and later, needs
testing on later macOS versions for any regressions.

Fixes issue #2377.