Cocoa: Fix modifier keys getting lost due to global system shortcuts
In normal situations modifier key release triggers flagsChanged with
the corresponding keyCode for that modifier key. If the user does a
global system shortcut such as screenshoting with cmd+ctrl+shift+4
then we will see the keyDown for the cmd+ctrl+shift but no keyUp or
flagsChanged after the screenshot is complete. This will leave
window->keys in the wrong state: the modifier keys will be stuck in
GLFW_PRESS until they are individually pressed and released again or
window focus changes.

When the screenshot chord is finished we _do_ get a flagsChanged event
but with the keyCode for "a" (which is weird) but the modifiers are
valid so we simply just emit each of their states and let _glfwInputKey
handle de-duplicating the event if needed.