Fixes an out of bounds joystick array access
Testing with an Xbox Series X controller. The extra button they
added directly under the Xbox button emits 0xA7 for me, which is
lower than BTN_MISC. As a result, when pressed, handleKeyEvent
was indexing an array with a negative index.

This commit adds a bounds check preventing the negative index, and
adds a second bounds check to a similar function preemptively.

It's possible that the better fix for this is to extend the range of
the buttons array so that this button can be mapped, but that may
need some disucssion first.