Cocoa: Fix glfwGetFramebufferSize() for non-retina windows.
On some Macs, glfwGetFramebufferSize() would return a high-resolution result even when the GL framebuffer was not created in high-resolution (retina) mode. This is because the screen-to-pixel conversion function, convertRectToBacking(), doesn't take into account whether the framebuffer was created in high-resolution mode; it will always return a high-resolution result if the device is high-res.

Now we check whether the framebuffer is indeed in retina mode before using the convertRectToBacking() function.

Fixes #2478